requirements.txt
openai>=1.12.0
python-pptx>=0.6.21
matplotlib>=3.7.0
plotly>=5.18.0
pandas>=2.0.0
pillow>=8.0.0
python-dotenv>=1.0.1
tenacity>=8.2.3
pytest>=8.0.0
pytest-asyncio>=0.23.5
aiohttp>=3.9.3
typing-extensions>=4.9.0
pytest-cov>=4.0.0  # For test coverage reporting
numpy>=1.24.0
nltk>=3.8.1
kaleido>=0.2.1  # Required for plotly static image export
pyyaml>=6.0.0  # For style import/export functionality 